Lots of modifications e.g. new configurations for sqldump and connection checks.
- Added new configurations to config.json. Possibility to enable or
disable sql dump and connection checks. Also, if those are enabled, it
is not possible to set the hour of the day sql dump is taken and also
possibility to set that what is the day for connection check 0-6 and
what is the hour that connection check is done.
- Sended connection checks are now also logged, so they are not sended
more than once / week. In some occasions those were sended many times.
- There was also bug in checking old entries from log that caused issue
when comparing current time to time when last warning was sended ->
sometimes warnings were not sended at all