Added insert command examples that use fahrenheits instead of celsius (at the end of the file)
Added insert command examples that use fahrenheits instead of celsius to
the end of the file. Comment current insert out and take insert that
uses fahrenheits to use if you want to use fahrenheits instead of
celsius. (This is just a quick way of changing...thinking about makin it
configurable on later updates)